requirements.txt
requests
pandas
numpy
tenacity
python-dotenv
matplotlib
plotly
yfinance
statsmodels
web3
py-clob-client
websockets
google-cloud-bigquery
db-dtypes
pyarrow
